ai-apps-collection/agno-apps/podcastbrain/requirements.txt
agno
google-genai
streamlit>=1.40.0
elevenlabs
python-dotenv>=1.0.0
pydantic>=2.0.0
